Action bar moved to LayoutUtils, some more pending...
